+++ NEWS.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.15.23]
+
+### Added
+
+* a field `point` to `ValidationFibreVector` to store potentially store the point of the vector.
+* a field `point` to `ValidationFibrePoint` to store potentially store the point of the point.
+### Changed
+
+* the internal function `_array_value` was renamed to `_value`,
+  since the cases where it is used also work for non-array poins/vectors
+
 ## [0.15.22] 15/11/2024
 
 ### Added